feat: implémenter le système initiale de session de jeu (timer, tour par tour...)

Voir la merge request : polytechnique-montr-al/log2990/20253/equipe-204/LOG2990-204!71

+++ server/app/modules/game-store/services/game-editor/game-editor.service.ts
@@ -70,11 +70,10 @@ export class GameEditorService {
 
     private async buildImageUpdate(dto: PatchGameEditorDto, id: string): Promise<Partial<GameDocument>> {
         if (!dto.gridPreviewUrl) return {};
-        
+
         const existingGame = await this.gameModel.findById(id).lean();
-        if (existingGame?.gridPreviewUrl)
-            await this.imageService.deleteImage(existingGame.gridPreviewUrl);
-        
+        if (existingGame?.gridPreviewUrl) await this.imageService.deleteImage(existingGame.gridPreviewUrl);
+
         const filename = `game-${id}-${Date.now()}-preview.png`;
         const gridPreviewUrl = await this.imageService.saveImage(dto.gridPreviewUrl, filename, 'game-previews');
         return { gridPreviewUrl };
